Revision 329:
- build.sh auf FloppyImage.img umgestellt (In Ermangelung von Linux kann ich dennoch nicht beurteilen, ob diese Datei funktioniert)
- HELLO.ELF jetzt in version 0.5.7 (Zählweise geändert, Float-Test weg, Wartet am Ende auf Tastendruck)
- Shell: testch nicht mehr genutzt
- Float-Test im ckernel.c weg, #NM-Ausgabe in Diagnosis-Modus verlegt
- "Aufräumarbeiten"

git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@329 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/time.c
@@ -22,20 +22,6 @@ tm_t* cmosTime(tm_t* ptm)
     return ptm;
 }
 
-
-///// CDI wrapper used. Needs to include cdi/cmos.h instead of cmos.h
-//tm_t* cmosTime(tm_t* ptm)
-//{
-//    ptm->second     = PackedBCD2Decimal(cdi_cmos_read(0x00));
-//    ptm->minute     = PackedBCD2Decimal(cdi_cmos_read(0x02));
-//    ptm->hour       = PackedBCD2Decimal(cdi_cmos_read(0x04));
-//    ptm->dayofmonth = PackedBCD2Decimal(cdi_cmos_read(0x07));
-//    ptm->month      = PackedBCD2Decimal(cdi_cmos_read(0x08));
-//    ptm->year       = PackedBCD2Decimal(cdi_cmos_read(0x09));
-//    ptm->century    = PackedBCD2Decimal(cdi_cmos_read(0x32));
-//    return ptm;
-//}
-
 static void appendInt(int val, char* dest, char* buf) {
     if (val<10)
     {
@@ -45,7 +31,7 @@ static void appendInt(int val, char* dest, char* buf) {
     strcat(dest, buf);
 }
 
-unsigned int calculateWeekday(unsigned int year, unsigned int month, unsigned int day) {
+unsigned int calculateWeekday(uint16_t year, uint8_t month, uint32_t day) {
     day += 6; //1.1.2000 was Saturday
     day += (year-2000)*365.25;
     if (month > 11)
